rename CQRS 패턴 파일 경로 수정
