+++ toyProject/src/main/java/com/mamully/toyProject/test/command/entities/Test.java
@@ -1,4 +1,4 @@
-package com.mamully.toyProject.test.entities;
+package com.mamully.toyProject.test.command.entities;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
