+++ toyProject/src/main/java/com/mamully/toyProject/test/command/repositories/TestRepository.java
@@ -1,6 +1,6 @@
-package com.mamully.toyProject.test.repositories;
+package com.mamully.toyProject.test.command.repositories;
 
-import com.mamully.toyProject.test.entities.Test;
+import com.mamully.toyProject.test.command.entities.Test;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TestRepository extends JpaRepository<Test, Long> {
