+++ toyProject/src/main/java/com/mamully/toyProject/test/command/services/TestDTO.java
@@ -1,4 +1,4 @@
-package com.mamully.toyProject.test.services;
+package com.mamully.toyProject.test.command.services;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
