+++ toyProject/src/main/resources/com/mamully/toyProject/test/query/mappers/TestMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.mamully.toyProject.test.mappers.TestMapper">
+<mapper namespace="com.mamully.toyProject.test.query.mappers.TestMapper">
     <!-- 모든 Test 데이터를 가져오는 SQL -->
     <select id="findAllTests" resultType="TestDTO">
         SELECT * FROM test
